+++ Documentation/CHANGELOG.md
@@ -11,3 +11,12 @@ Added:
 - Achievements and Leaderboard integration
 - Power-ups (Health, Shield) and pick-ups that activate missile launchers on the map.
 
+### [0.6.0] - 2023-09-26
+Polished version.
+
+Added:
+- New VR enviroment
+- Ability to adjust table height
+- Virtual joy stick on hand tracking mode
+- New foot mesh
+- Minor bug fixes